Move fare services into their separate package
