+++ src/main/java/org/opentripplanner/routing/fares/AddingMultipleFareService.java
@@ -5,7 +5,6 @@ import org.opentripplanner.routing.algorithm.raptor.transit.TripSchedule;
 import org.opentripplanner.routing.core.Fare;
 import org.opentripplanner.routing.core.Fare.FareType;
 import org.opentripplanner.routing.core.Money;
-import org.opentripplanner.routing.services.FareService;
 import org.opentripplanner.transit.raptor.api.path.Path;
 
 import java.io.Serializable;
