+++ src/main/java/org/opentripplanner/routing/fares/FareService.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.services;
+package org.opentripplanner.routing.fares;
 
 import org.opentripplanner.routing.algorithm.raptor.transit.TransitLayer;
 import org.opentripplanner.routing.algorithm.raptor.transit.TripSchedule;
