+++ src/main/java/org/opentripplanner/routing/fares/impl/Ride.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.impl;
+package org.opentripplanner.routing.fares.impl;
 
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Stop;
