+++ src/main/java/org/opentripplanner/routing/fares/impl/RideMapper.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.impl;
+package org.opentripplanner.routing.fares.impl;
 
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.TripPattern;
