+++ src/main/java/org/opentripplanner/routing/fares/impl/SFBayFareServiceFactory.java
@@ -1,6 +1,6 @@
-package org.opentripplanner.routing.impl;
+package org.opentripplanner.routing.fares.impl;
 
-import org.opentripplanner.routing.services.FareService;
+import org.opentripplanner.routing.fares.FareService;
 
 public class SFBayFareServiceFactory extends DefaultFareServiceFactory {
     @Override
