+++ src/main/java/org/opentripplanner/routing/vehicle_rental/TimeBasedVehicleRentalFareService.java
@@ -7,7 +7,7 @@ import org.opentripplanner.routing.algorithm.raptor.transit.TripSchedule;
 import org.opentripplanner.routing.core.Fare;
 import org.opentripplanner.routing.core.Fare.FareType;
 import org.opentripplanner.routing.core.WrappedCurrency;
-import org.opentripplanner.routing.services.FareService;
+import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.transit.raptor.api.path.PathLeg;
 
 import java.io.Serializable;
