+++ src/main/java/org/opentripplanner/routing/vehicle_rental/TimeBasedVehicleRentalFareServiceFactory.java
@@ -10,8 +10,8 @@ import java.util.Map;
 
 import org.opentripplanner.model.OtpTransitService;
 import org.opentripplanner.common.model.P2;
-import org.opentripplanner.routing.services.FareService;
-import org.opentripplanner.routing.services.FareServiceFactory;
+import org.opentripplanner.routing.fares.FareService;
+import org.opentripplanner.routing.fares.FareServiceFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
