+++ src/test/java/org/opentripplanner/routing/algorithm/TestFares.java
@@ -15,8 +15,8 @@ import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.core.WrappedCurrency;
 import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.impl.SeattleFareServiceFactory;
-import org.opentripplanner.routing.services.FareService;
+import org.opentripplanner.routing.fares.impl.SeattleFareServiceFactory;
+import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.routing.spt.GraphPath;
 import org.opentripplanner.routing.spt.ShortestPathTree;
 import org.opentripplanner.util.TestUtils;
